+++ apps/api/src/pkg/db.ts
@@ -2,12 +2,15 @@ import { type PlanetScaleDatabase, drizzle } from "drizzle-orm/planetscale-serve
 
 import { Client } from "@planetscale/database";
 import { schema } from "@unkey/db";
+import type { Logger } from "@unkey/worker-logging";
 export type Database = PlanetScaleDatabase<typeof schema>;
 
 type ConnectionOptions = {
   host: string;
   username: string;
   password: string;
+  retry: number | false;
+  logger?: Logger;
 };
 
 export function createConnection(opts: ConnectionOptions): Database {
@@ -24,7 +27,24 @@ export function createConnection(opts: ConnectionOptions): Database {
         if (u.host.includes("localhost")) {
           u.protocol = "http";
         }
-        return fetch(u, init);
+
+        if (!opts.retry) {
+          return fetch(u, init);
+        }
+
+        let err: Error | undefined = undefined;
+        for (let i = 0; i <= opts.retry; i++) {
+          try {
+            return fetch(u, init);
+          } catch (e) {
+            opts.logger?.warn("fetching from planetscale failed", {
+              url: u.toString(),
+              attempt: i + 1,
+            });
+            err = e as Error;
+          }
+        }
+        throw err;
       },
     }),
     {