+++ apps/semantic-cache/src/pkg/hono/app.ts
@@ -1,18 +1,11 @@
-import { Hono } from "hono";
+import { type Context as GenericContext, Hono } from "hono";
 import type { HonoEnv } from "./env";
 
 export function newApp() {
   const app = new Hono<HonoEnv>();
 
   app.use("*", (c, next) => {
-    c.set(
-      "location",
-      c.req.header("True-Client-IP") ??
-        c.req.header("CF-Connecting-IP") ??
-        // @ts-expect-error - the cf object will be there on cloudflare
-        c.req.raw?.cf?.colo ??
-        "",
-    );
+    c.set("location", c.req.header("True-Client-IP") ?? c.req.header("CF-Connecting-IP") ?? "");
     c.set("userAgent", c.req.header("User-Agent"));
 
     return next();
@@ -22,3 +15,5 @@ export function newApp() {
 }
 
 export type App = ReturnType<typeof newApp>;
+
+export type Context = GenericContext<HonoEnv>;