+++ internal/checkly/src/__checks__/api/ratelimit/group.ts
@@ -1,0 +1,7 @@
+import { CheckGroup } from "checkly/constructs";
+import { ALL_LOCATIONS } from "../../../locations";
+
+export const ratelimitsV1 = new CheckGroup("/v1/ratelimits", {
+  name: "ratelimits",
+  locations: ALL_LOCATIONS,
+});