+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @unkey/web
 
+## 0.1.26
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.12
+
+## 0.1.25
+
+### Patch Changes
+
+- @unkey/ratelimit@0.1.11
+
 ## 0.1.24
 
 ### Patch Changes