+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @unkey/nuxt
 
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [69c88fa]
+  - @unkey/api@0.20.7
+
+## 1.1.14
+
+### Patch Changes
+
+- Updated dependencies [56ad960]
+  - @unkey/api@0.20.6
+
 ## 1.1.13
 
 ### Patch Changes